+++ packages/components/src/Menu/index.ts
@@ -1,4 +1,6 @@
 export * from './Menu'
+export * from './MenuDisclosure'
+export * from './MenuList'
 export * from './MenuGroup'
 export * from './MenuItem'
 export * from './MenuSearch'