+++ packages/components/src/Overlay/useFocusTrap.hook.ts
@@ -1,5 +1,5 @@
 import createFocusTrap, { FocusTrap } from 'focus-trap'
-import { MutableRefObject, useEffect, useRef } from 'react'
+import { useRef, useCallback } from 'react'
 
 export interface UseFocusOptions {
   escapeDeactivates?: boolean
@@ -9,29 +9,27 @@ export interface UseFocusOptions {
 export function useFocusTrap(
   escapeDeactivates = false,
   clickOutsideDeactivates = true
-): MutableRefObject<HTMLDivElement | null> {
+) {
   const trap = useRef<FocusTrap>()
-  const ref = useRef<HTMLDivElement>(null)
 
-  useEffect(() => {
-    if (ref.current === null) {
-      trap.current && trap.current.deactivate()
-    } else {
-      const autoFocusElement = ref.current.querySelector(
-        '[data-autofocus="true"]'
-      ) as HTMLElement
-      trap.current = createFocusTrap(ref.current, {
-        clickOutsideDeactivates,
-        escapeDeactivates,
-        fallbackFocus: ref.current,
-        ...(autoFocusElement ? { initialFocus: autoFocusElement } : {}),
-      })
+  return useCallback(
+    node => {
+      if (!node) {
+        trap.current && trap.current.deactivate()
+      } else {
+        const autoFocusElement = node.querySelector(
+          '[data-autofocus="true"]'
+        ) as HTMLElement
+        trap.current = createFocusTrap(node, {
+          clickOutsideDeactivates,
+          escapeDeactivates,
+          fallbackFocus: node,
+          ...(autoFocusElement ? { initialFocus: autoFocusElement } : {}),
+        })
 
-      trap.current.activate()
-    }
-    return () => {
-      trap.current && trap.current.deactivate()
-    }
-  }, [escapeDeactivates, clickOutsideDeactivates])
-  return ref
+        trap.current.activate()
+      }
+    },
+    [escapeDeactivates, clickOutsideDeactivates]
+  )
 }